[detection] Optim of the outlierRemoval function
Consider a subset of the input points during the robust ellipse
estimation. All the inliers are then added at the end of the procdedure,
i.e. those whose the distance to the estimated ellipse is then than the
computed LMDeS threshold.

The idea is simply to avoid to compute many hundreds of point-to-ellipse
distances at each iteration of the algorithm.