use symbol for list size counters (1 config)
Allocating a pinned host memory page for every int that is a list size
counter is annoying. Without pinned, every int transfer is blocking.
Apparently symbols do not have the limitation. However, that does not
work for CUB functions, so we need 1 device-sided int per frame. But
we can copy device-to-symbol and symbol-to-host asynchronous. Hope that
works.