[detection] Fix the mess with the _processed attribute.
Now, that attribute takes the value of the index of the candidate over
the loop two. In order to avoid to reset it, it is incremented and at
any iteration, an edge has been processed within the same iteration is
the value of _processed is the same as runId, otherwise the edge point
can be added to the outerEllipsePoints.